services/catalog/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.6.1
motor==3.6.0
pymongo==4.10.1
